requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
google-generativeai==0.3.2
pydantic==1.10.12
PyPDF2==3.0.1
python-docx==0.8.11
mammoth==1.6.0
beautifulsoup4==4.12.2
langchain==0.1.0
langchain-google-genai==0.0.5
psutil==5.9.8
cloudinary==1.36.0
# faiss-cpu==1.7.4  # Commented out due to compilation issues
